requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlmodel==0.0.14
sqlalchemy==2.0.25

# Milvus Vector Database
pymilvus==2.3.4

# LangChain & LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langsmith>=0.1.0
langchain-text-splitters>=0.3.0

# GLM AI & Embeddings
zhipuai>=2.1.5
openai>=1.10.0
tiktoken

# Tools and Search
tavily-python>=0.3.0
httpx>=0.26.0
aiohttp>=3.9.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
loguru==0.7.2
python-dateutil==2.8.2
pydantic[email]==2.5.3
python-docx==1.1.0
docx2txt==0.8
pypdf==4.0.1

# Development Tools
black==24.1.1
mypy==1.8.0
isort==5.13.2
pytest==7.4.4
pytest-asyncio==0.23.3
